Move comments to a url hover bar.
This does a couple things. First, it moves the comment section of yaml
files to a hoverable title on the link, so screen space isn't wasted.

But, users don't know it's there, and it throws out any markup.

I find it's styled much nicer this way on smaller devices; otherwise,
the responsive tables all turn into one long line.

It'll work for now and tbh, I'm not a fan of accepting markup in YAML.